requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

confection==0.1.5
hydra-core==1.3.2
lightning==2.4.0
matplotlib==3.9.2
numpy==2.0.2
pillow==10.4.0
spacy==3.8.2
torch==2.4.1
tqdm==4.66.5
